Fix import.

+++ batchglm/train/numpy/glm_nb/estimator.py
@@ -83,8 +83,7 @@ class Estimator(EstimatorGlm):
             input_data=input_data,
             model_vars=self.model_vars,
             compute_mu=self._train_loc,
-            compute_r=not self._train_scale,
-            dtype=dtype,
+            compute_r=not self._train_scale
         )
         super(Estimator, self).__init__(input_data=input_data, model=model, dtype=dtype)
